refactor Language and LanguageManage into /config/language
