+++ app/src/main/java/com/github/fribourgsdp/radio/MyAppCompatActivity.kt
@@ -2,6 +2,7 @@ package com.github.fribourgsdp.radio
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.github.fribourgsdp.radio.config.language.LanguageManager
 
 open class MyAppCompatActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
